Moved Resilience methods/properties to partial class
